Ensure color swatches wrap inside the product tile
With this change we'll ensure that the color swatches inside
the product tile wrap properly, allowing the customers to pick
a swatch when the product includes a lot of colorways.
